backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings